+++ shared/src/__tests__/SfProgress/SfProgressBase.spec.js
@@ -1,0 +1,13 @@
+import { shallowMount } from '@vue/test-utils'
+import SfProgressBase from '../../components/SfProgress/SfProgressBase.vue'
+
+let wrapper
+
+describe('SfProgressBase.vue', () => {
+  beforeEach(() => {
+    wrapper = shallowMount(SfProgressBase)
+  })
+  it('render component correctly', () => {
+    expect(wrapper.html()).toMatchSnapshot()
+  })
+})