+++ shared/src/components/SfTag/SfTagIconRight.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="absolute flex px-3 py-1.5 text-sm font-normal text-gray-600 bg-gray-100 rounded-md font-body direction-row">
+    <slot>
+      Label
+    </slot>
+    <div class="w-4 h-4 ml-2 bg-black rounded-full" />
+  </div>
+</template>
+
+<script>
+export default {
+  name: "SfTagIconRight"
+}
+</script>