+++ shared/docs/v0.0.1/changelog.md
@@ -2,6 +2,9 @@
 
 ## 🚀 Features:
 - add VueUse package
+- Spinner component created with stories and tests 
+- Tag component created with tests and stories 
+- Sheet component created
 
 ## 🐛 Fixes: